+++ src/cli/mod.rs
@@ -1,3 +1,4 @@
 pub(crate) mod command;
 pub(crate) mod decode;
+pub(crate) mod evo;
 pub(crate) mod info;